Move PluginClassInfo related to shared helper
